Now the game is drawn only when it changes. Main loop logic is now based on time, rather than on inaccurate "ticks"

+++ source/gameCore.h
@@ -26,7 +26,7 @@ private:
 	graphics graphicsObj;
 	sound soundObj;
 	entity* player;
-	long tick;
+	long long tick;
 	void createSavefile(string saveName);
 	void loadSavefile(string saveName);
 	void gameLaunch();
